[s3] Fix downloading from S3
Summary:
This script now performs the unscalable action of keeping all S3 keys in
memeory. For a project of this size/scope, this is acceptable.

Description of bug being fixed:
boto's `get_all_data` method only returns 1000 keys from a bucket. Use the list
method to page through all of the keys.